story-on-board-github/pyproject.toml
[project]
name = "story-on-board"
version = "1.0.0"
description = "AI-powered multimodal storybook and storyboard creator"
authors = [
    {name = "Story on Board Team"}
]
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "websockets==14.1",
    "google-generativeai==0.8.3",
    "google-cloud-aiplatform==1.71.1",
    "google-cloud-texttospeech==2.34.0",
    "vertexai==1.71.1",
    "pillow==11.1.0",
]

[project.urls]
Homepage = "https://story-on-board-474322067410.us-central1.run.app"
Repository = "https://github.com/YOUR_USERNAME/story-on-board"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
